refactor(auth-service): divide controller class to become more moduler
