+++ backend/services/auth-service/src/main/java/com/flightroster/auth/controllers/AuthController.java
@@ -19,7 +19,9 @@ import com.flightroster.auth.services.AuthService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
-// ===== Class =====
+// ===========================================
+// 					Controllers 
+// ===========================================
 
 @RestController
 @RequestMapping("/auth")
@@ -33,11 +35,11 @@ public class AuthController
 	// ================================
 
 	/**
-	 * @brief Registers a new user.
-     * 
-	 * @route POST /auth/register
-     * 
-	 * @access Public
+	 * Registers a new user.
+	 * 
+	 * route: POST /auth/register
+	 * 
+	 * access: Public
 	 */
 	@PostMapping("/register")
 	public ResponseEntity<AuthResponse> register(@Valid @RequestBody RegisterRequest request)
@@ -46,11 +48,11 @@ public class AuthController
 	}
 
 	/**
-	 * @brief Authenticates user credentials and issues tokens.
+	 * Authenticates user credentials and issues tokens.
      * 
-	 * @route POST /auth/login
+	 * route: POST /auth/login
      * 
-	 * @access Public
+	 * access: Public
 	 */
 	@PostMapping("/login")
 	public ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest request)
@@ -59,38 +61,37 @@ public class AuthController
 	}
 
 	/**
-	 * @brief Health check endpoint for monitoring service status.
+	 * Health check endpoint for monitoring service status.
      * 
-	 * @route GET /auth/health
+	 * route: GET /auth/health
      * 
-	 * @access Public
+	 * access: Public
 	 */
 	@GetMapping("/health")
-	public ResponseEntity<String> health()
+	public ResponseEntity<Object> health()
 	{
-		return (ResponseEntity.ok("Auth service is running"));
+		return (ResponseEntity.ok(java.util.Map.of("status", "Auth service is healthy")));
 	}
 
 	/**
-	 * @brief Provides public JSON Web Key Set (JWKS) for JWT validation.
+	 * Refreshes access token using a valid refresh token.
      * 
-	 * @route GET /.well-known/jwks.json
+	 * route: POST /auth/refresh-token
      * 
-	 * @access Public
+	 * access: Public
 	 */
-	@GetMapping("/.well-known/jwks.json")
-	public ResponseEntity<String> jwks()
+	@PostMapping("/refresh-token")
+	public ResponseEntity<AuthResponse> refreshToken(@Valid @RequestBody RefreshTokenRequest request)
 	{
-		// Placeholder for JWKS implementation
-		return (ResponseEntity.ok("{\"keys\": []}"));
+		return (ResponseEntity.ok(authService.refreshAccessToken(request)));
 	}
 
 	/**
-	 * @brief Handles password reset initiation (sending email, etc).
+	 * Handles password reset initiation (sending email, etc).
      * 
-	 * @route POST /auth/forgot-password
+	 * route: POST /auth/forgot-password
      * 
-	 * @access Public
+	 * access: Public
 	 */
 	@PostMapping("/forgot-password")
 	public ResponseEntity<String> forgotPassword()
@@ -113,36 +114,26 @@ public class AuthController
 		return (ResponseEntity.ok("Password reset successful (mock)"));
 	}
 
-	// ================================
-	// PROTECTED ENDPOINTS
-	// ================================
 
 	/**
-	 * @brief Logs out the current user and revokes their refresh token.
+	 * Logs out the current user and revokes their refresh token.
      * 
 	 * @route POST /auth/logout
      * 
-	 * @access Protected
+	 * @access Public
 	 */
 	@PostMapping("/logout")
-	public ResponseEntity<String> logout()
+	public ResponseEntity<Object> logout(@Valid @RequestBody RefreshTokenRequest request)
 	{
-		// TODO: Implement logout (revoke refresh token)
-		return (ResponseEntity.ok("Logged out successfully"));
+		authService.logout(request);
+
+		return (ResponseEntity.ok(java.util.Map.of("message", "Logged out successfully")));
 	}
 
-	/**
-	 * @brief Refreshes access token using a valid refresh token.
-     * 
-	 * @route POST /auth/refresh-token
-     * 
-	 * @access Protected
-	 */
-	@PostMapping("/refresh-token")
-	public ResponseEntity<AuthResponse> refreshToken(@Valid @RequestBody RefreshTokenRequest request)
-	{
-		return (ResponseEntity.ok(authService.refreshAccessToken(request)));
-	}
+
+	// ================================
+	// PROTECTED ENDPOINTS
+	// ================================
 
 	/**
 	 * @brief Validates the provided JWT access token.
@@ -184,50 +175,5 @@ public class AuthController
 	{
 		// TODO: Implement password change logic
 		return (ResponseEntity.ok("Password changed successfully"));
-	}
-
-	// ================================
-	// INTERNAL ENDPOINTS
-	// ================================
-
-	/**
-	 * @brief Validates tokens for inter-service communication.
-     * 
-	 * @route POST /auth/internal/validate-token
-     * 
-	 * @access Internal (requires ADMIN)
-	 */
-	@PostMapping("/internal/validate-token")
-	public ResponseEntity<String> validateInternalToken()
-	{
-		return (ResponseEntity.ok("Internal token validated"));
-	}
-
-	/**
-	 * @brief Disables a user account.
-     * 
-	 * @route POST /auth/internal/disable-user
-     * 
-	 * @access Internal (requires ADMIN)
-	 */
-	@PostMapping("/internal/disable-user")
-	public ResponseEntity<String> disableUser()
-	{
-		// TODO: Implement disable logic
-		return (ResponseEntity.ok("User disabled successfully"));
-	}
-
-	/**
-	 * @brief Synchronizes a user's role with another service.
-     * 
-	 * @route PATCH /auth/internal/sync-role
-     * 
-	 * @access Internal (requires ADMIN)
-	 */
-	@PatchMapping("/internal/sync-role")
-	public ResponseEntity<String> syncRole()
-	{
-		// TODO: Implement role synchronization
-		return (ResponseEntity.ok("User role synchronized successfully"));
 	}
 }
